requirements.txt
strands-agents[ollama]==1.18.0
strands-agents-tools==0.2.16
fastapi==0.122.0
uvicorn==0.38.0
python-dotenv==1.2.1